Adds support for using an arbitrary branch to trigger releases
Previously, the default_branch would match the CloudWatch Event,
which would trigger the build, but the branch value wasn't
available to the job so it always executed on the actual default
branch.

This patch passes the value through the CodeBuild environment, and
checks out the specified branch prior to running the versioning
commands. This allows this module to use any arbitrary branch to
trigger a release.

The var name is changed from default_branch to release_branch to
better indicate the capability.