Changes count to reference same var as linked resource
Fixes errors like this:

```
Error: Invalid count argument

  on ../../modules/_internal/runner/main.tf line 116, in resource "aws_iam_role_policy_attachment" "codebuild":
 116:   count = length(data.aws_iam_policy.codebuild.*.arn)

The "count" value depends on resource attributes that cannot be determined
until apply, so Terraform cannot predict how many instances will be created.
To work around this, use the -target argument to first apply only the
resources that the count depends on.
```